+++ piecrust/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Implemented Display for ModuleId. [#178]
 - Removed SELF_ID export from contracts. [#167]
 - Added uplink::owner and uplink::self_id. [#158]
 - Added persistence fo module metadata. [#167]
@@ -32,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#158]: https://github.com/dusk-network/piecrust/issues/158
 [#162]: https://github.com/dusk-network/piecrust/issues/162
 [#167]: https://github.com/dusk-network/piecrust/issues/167
+[#178]: https://github.com/dusk-network/piecrust/issues/178
 
 <!-- VERSIONS -->
 [Unreleased]: https://github.com/dusk-network/piecrust/compare/v0.1.0...HEAD